Remove script used for authorizing changes to user
Originally script used for quick and 'dirty' authorization. After
using Helmet to set response headers this script was blocked
from executing so I removed it.

Could've used a nonce to enable execution but not really an issue
as next step in the project is to use another middleware to handle
the authentication and sessions for the app.